remove obscure directory mismatch check
Summary:
this is an ancient check in which `flow status` confirms that the server and client both detected the same root directory, even though they're talking to each other over a socket that's dependent on that being true.

the only known way this mismatch can be hit is because there is a bug where `cd /foo; flow --flowconfig-name bar/.myflowconfig` starts a server in `/foo` with the socket `/tmp/flow/zSfoozSbarzS.myflowconfig.sock`, and `cd /foo/bar; flow --flowconfig-name .myflowconfig` starts a server in `/foo/bar` but uses the same socket.

obviously we should fix that, but (a) it's *so* obscure, and (b) this applies to *only* `flow status`, so isn't entirely effective anyway. I'd like to reap this check now because of the invasive place it lives in `CommandHandler.handle_parallelizable_ephemeral_unsafe`

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D32304175

fbshipit-source-id: cb0608a03e1f90f203e87d7915525c9a0006c2b4