Fix shorthands in ast mapper
Summary:
This diff undoes the change to `shorthand` in the ast mapper in D31453075 (https://github.com/facebook/flow/commit/806cc9c6c3023cf4c1a82bb4114aa5239fd3285e), which caused issues for prettier (https://github.com/facebook/flow/issues/8776). This change is important for codemods, because a codemod can change the key to an object property without changing the value, so this diff removes that behavior from the base visitor and adds it to the codemod visitor.

Fixes #8776

Changelog: [internal]

Reviewed By: mroch

Differential Revision: D32258284

fbshipit-source-id: fd2f8594bd1ca5d33bc626dc6fb86445a87df1b7