Change parser to require annotations on `declare`d variables
Summary:
Currently it's allowed to create declared variables without annotations:

```
declare var foo;
```
When this happens, `foo`'s type is an uninhabited tvar, and since the variable likely will never receive lower bounds its likely to effectively be `empty`.

This is, of course, incompatible with local type inference. To prevent this, this diff alters the parser to require that all `declare`d variables be annotated.

Changelog: [errors] The Flow parser now requires that variables declared using the `declare var` syntax be annotated with a type.

Reviewed By: mroch

Differential Revision: D32214530

fbshipit-source-id: 75549e82819dcf99ca8b649d425c45e13934fcbd