Make the name_resolver and ssa_builder handle prefixed update expressions, op-assign correctly
Summary:
Recall that `++x` and `x++` are different in that the value of the overall expression is the incremeneted `x` in the first place and the original `x` in the second place. The SSA builder and name_resolver weren't handling the prefix case--whether the update is prefixed or not, they would see the read of `x` going to the original definition of `x`. This fixes.

ALSO FIXES SOME STUFF WITH OP ASSIGNNN

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D32215470

fbshipit-source-id: 534a3fbb82676c71dbd204093d3c713c18f829e0