[insert-type] fix case of importing an ES6 default export
Summary:
Services that insert type imports lookup the type_sig of the file defining the name
to be imported to determine two things: (i) whether the name is exported, and (ii) the
kind of import (default or not).

The form `export default class C {}` was not supported before (the import insertion
would always fail) because we were not looking for `ExportDefaultBinding` in
insert_type_imports.ml. This diff fixes this by checking whether the binding that
corresponds to `ExportDefaultBinding` has the same name as the one we're
querying for. If so, then we add a default import for this name (`import type x ...`).

Changelog: [fix] Fixed a bug that prevented code-action/codemod services from adding type imports for `default export class` definitions

Reviewed By: samwgoldman

Differential Revision: D32378155

fbshipit-source-id: 6567b4eb2896b529e90cfebd54335d9486e00808