functorize Destructuring over Statement & don't pass `expr` parameter
Summary:
As I started working on passing down "expected" types for contextual typing, I
noticed that the `expr` parameter passed to Destructuring functions was missing
the `~annot` parameter. Instead of just adding it and further complicating the
module's API, I saw a chance to cleanup the Destructuring API.

Now that modules in recursive dependency with statement.ml are parameterized
by Env, we can selectively extend their functor parameters to simplify their APIs:
This way Destructuring doesn't need to pass around `~expr`, it can call directly to
Statement.expression.

IMO this is also a cleaner way to determine the true dependencies of this module.

This change required a couple more changes:
* Since Func_stmt_config depends on Destructuring, I pulled it out in its own file.
* Type_inference_js.Make now needs to tie the knot of these recursive modules.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D32401622

fbshipit-source-id: 31162b242177cc14c1341a944a46f9e6aa37cfc3