fix for direct_dependent_files computation when force-checking files
Summary:
consider two files, `a.js` and `b.js` where `a` depend on `b`. we delete `b.js` and then `flow force-recheck --focus a.js b.js`. result:

```
new_or_changed = []
deleted = ['b.js']
unchanged = ['a.js']
files_to_force = ['a.js', 'b.js']
unchanged_files_to_force = ['a.js']
unchanged_files_with_dependents = ['a.js']
```
we need to re-resolve the requires in `a.js` to realize that it no longer points at `b.js`.

however, we don't do anything in `commit_modules_and_resolve_requires`, because `a.js` is unchanged.

in the `direct_dependent_files` computation, we have:
```
candidates = unchanged - unchanged_files_with_dependents = []
root_files = ['a.js']
root_modules = ['a']
direct_dependent_files (result) = []
```
but we need `direct_dependent_files` to include `a` so that the next step, "ReresolveDirectDependents", will reresolve `a`.

notably, `candidates` is empty, even though `a.js` is a direct dependent of a deleted file.

the comments on `Dep_service.calc_direct_dependents` say that `candidates` and `root_files` *should* be disjoint, and that's not the case with this diff, in the case of unchanged force-focused files. this doesn't seem to be an actual invariant of the algorithm and I think this is safe because including more direct dependent files is more conservative. and in my next diff, I can remove the `unchanged_files_with_dependents` from `root_files` which restores this invariant anyway.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D32454735

fbshipit-source-id: 20d317efe7eb2609876553b2d2b8ad687854af60