[relay-integration] Add option to exclude paths from Relay integration
Summary:
This diff adds a `relay_integration.excludes` option that works similarly to our other such exclude lists. You can use it to define multiple patterns for which you do not want to enable the Relay integration option.

Context:
When enabled, the Relay integration option treats GraphQL literals as requires to the Relay generated modules for that literal.

Reviewed By: mroch

Differential Revision: D32485175

fbshipit-source-id: 80f5f697f81865f77168fb498f1fb67564b043c0