[annot-inf] wrap all resolved constraints in elab_open in lazy constraints
Summary:
Before, definitions like
```
declare var x: {
  p: number;
  q: typeof (x.p);
};
```
would cause an [invalid-recursive-exported-annotation] error on the recurisve reference
to `x`. This is the result of a check in `ensure_annot_resolved` that ensures we are
not attempting to force into evaluation recursively defined lazy values (which would
otherwise be a runtime failure). In this example, annotation inference reaches this recursion
while attempting to recursively compute types for `x` -> `x.p` -> `x`.

The lazy constraint is created in `mk_tvar_lazy` and is not expected to be forced until
the entire body of the type has been visited. However, `elab_open` ended up forcing this
constraint to compute `x.p`, since `find_avar` returns `Annot_resolved` for the
recursive instance of the type for `x`.

This change prevents this by introducing a lazy indirection in the result of applying
`op` in `elab_open`. Laziness in the result of `x.p` delays forcing the type of `x` until
we're done visiting the object type.

Reviewed By: samwgoldman

Differential Revision: D32500366

fbshipit-source-id: 524072ffa055affb2b1536112938d62cd49937c9