Ordering for refinements that depend on a separate name
Summary:
This diff makes the reordering system aware of refinements. For the most part, refinements don't affect the ordering of a program--we already know thanks to the name_resolver what the refinement to be applied to a variable is, so we can resolve the variable before actually checking the expression that introduces the refinement. However, in cases like `instanceof`, we need to know the type of the RHS of the refinement before we can actually apply it. This diff extends the ordering system to ensure that such RHSs are checked before any refined variables that depend on them.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D32181857

fbshipit-source-id: 1096667f21b049ea7202b0feb15626ecb4740dab