Defs and ordering of class bindings
Summary:
This extends the name_def and name_def_ordering analyses to handle classes.

Classes are special with respect to cycles, because we'll assume that the type of a class that is fully annotated can be resolved before actually checking the class. Classes can also be used as either terms or types, and because of that duality the `Key` data structure that was previously used to distinguish type and term bindings doesn't work well anymore. Now, the analysis is over raw locations, and when we need to determine whether a binding is type-like or term-like, we just look at the `def` node that the binding is defined by and see what kind it is.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D32150064

fbshipit-source-id: e71d486fbb468361376cc2312ff0857fd3ee5f91