Create UninitializedClass write_state for classes before their declaration
Summary:
The duality of classes is mysterious.

When viewed as values, they're just let-bound, and should be a TDZ error if read from before their definition.
However, their types are hoisted, and so a type-level reference to a class should succeed.

This is now modeled with a new `write_state`, `UninitializedClass`, which acts as `Uninitialized` before a class is declared when referred to as a value and `Loc` when referred to as a type. This also means we now need to specify if we're querying for a type or a value (which we eventually would have to do anyways).

When class bindings are put in scope in the name_resolver, they start out as `UninitializedClass`es rather than `Uninitialized` like normal variables.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D32152956

fbshipit-source-id: 5b4cb375389578f5e4251295bf67097f57a35df1