fix rechecks being unnecessarily canceled due to deleted files
Summary:
during a recheck, the set of files that we are currently rechecking is kept in a mutable `will_be_checked_files`. it's used to ensure that future updates don't cancel a recheck to recheck the same files that are already in progress.

we set it to `files_to_force` when starting the recheck, and later update it to `to_merge_or_check`, which includes the dependencies and dependents that we determined also need to be included.

HOWEVER, deleted files are unique because deleting a file triggers a recheck -- in order to update the dependency graph and recheck dependents -- but deleted files can't themselves be rechecked, so they are not in `to_merge_or_check`.

so, when we REPLACE `will_be_checked_files` with `to_merge_or_check` instead of just ADDING `to_merge_or_check`, we remove the deleted files. then any command with a deleted unchecked dependency will cancel the recheck incorrectly.

this leads to enormous recheck series. pretty much the only way to get it to complete is to leave the IDE idle so as to avoid any commands that would restart it again. :-O

Changelog: [fix] Fix a bug that caused rechecks to be unnecessarily invalidated and restarted

Reviewed By: jbrown215

Differential Revision: D32554424

fbshipit-source-id: ca055b3028a615c6ac3e8036209a88f08581136d