Add file_key to resolved requires
Summary:
The file key is used as the hash table key for these heap values, but when
iterating over all resolved requires, we don't necessarily know the key already.
This field is necessary to reverse dependencies when iterating (see later diff
in stack for use).

I don't expect that this will cost much heap space, since most of the file key
will be duplicated in the other paths, so compression should be highly
effective.

Changelog: [internal]

Reviewed By: mroch

Differential Revision: D32600236

fbshipit-source-id: 36e31812086e2bb1beb9c81ab27d9b4de5f03032