Optimize calc_direct_dependents
Summary:
Before this change, calc_direct_dependents had a very inefficient
implementation, which had two parts. The first part extracts the entire
dependency graph in workers and returns it to the server process. The second
part traverses the dependency graph to accumulate dependents.

This is slow for many reasons:
* The first part iterates over all unchanged files, making it O(repo)
* The workers send back a large amount of data to the server
* Reversing the dependecy graph is O(repo) and blocks the server process

This diff takes another approach: We pass the set of changed files and modules
to workers, each of which reads the entire dependency graph from shared memory,
which we can do with the new Module_heaps.iter_resolved_requires API.

This solves each of the problems above:
* The parallel phase is O(changes)
* Workers only send back dependents which match the changes
* The server process only needs to merge sets of dependents

Changelog: [perf] Optimize dependent files calculation

Reviewed By: mroch

Differential Revision: D32609360

fbshipit-source-id: 21034aa93d4554944da6694329893845be8f9e22