prioritize more unchecked deps
Summary:
we're not always cancelling a normal recheck, depending on when during the recheck we try to check-contents, because of when `will_be_checked_files` gets updated.

### background

`will_be_checked_files` is a ref passed throughout `Types_js.recheck`, which is updated as the recheck figures out which additional files need to be included. for example, if rechecking A needs to check dependency B and dependent C, then `will_be_checked_files` will start with `A` and be updated to include `B` and `C` after `include_dependencies_and_dependents` figures that out.

`will_be_checked_files` is read by the `get_forced` function passed to `run_but_cancel_on_file_changes` during a recheck, which in turn passes it to `recheck_fetch`. `recheck_fetch` processes updates from the recheck stream, such as watchman events and unchecked dependency requests from the IDE commands. it uses `get_forced` (and therefore `will_be_checked_files`) to ignore updates that are already in progress. *only* if it gets an update that isn't ignored does it cancel the recheck and start a new one including that update.

### bug

the problem is that, when we receive a `DependenciesToPrioritize` workload, we want to cancel a normal recheck no matter what it's working on, but we *don't* want to cancel a priority recheck that's already working on the same files.

we get the latter part right, but we ignore the dependencies to prioritize if they're already part of a normal recheck. we should be canceling the normal recheck to prioritize them instead!

NOTE: we shouldn't be cancelling priority rechecks when we get more unchecked deps! they're independent from each other, so we should just finish and then do the next one. that's coming in future diffs.

### fix

the key here is knowing whether we're currently in a normal or a priority recheck. if a priority recheck, then filter unchecked dependency requests against the already-prioritized files in `will_be_checked_files` (via `get_forced`). if a normal recheck, ignore `will_be_checked_files` because we want to prioritize them instead.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D32636397

fbshipit-source-id: f9adcd34a1ec364b0bfcd75424ab692c246edd79