New type sig encoding
Summary:
(I apologize for the big diff. The important changes are in parser_utils, where
the new encoding is defined. The rest is reddiff or mechanical changes.)

Before this diff, we stored type sigs in shared memory using a the shared heap's
value encoding. That is, signature data was encoded as heap objects, each with
their own tag. The pointers within a signature were encoded as heap addresses.
Garbage collection would need to visit and mark these objects and move them
individually when compacting.

However, signatures are self contained. That is, nothing points directly to
signature data except the signature itself. It is not useful to have an
individual type export be addressible from anywhere in the heap, but we do pay a
cost: mark/sweep and compaction need to scan more addresses, compaction needs
to move more, smaller values.

This diff replaces the storage for signatures with a new encoding. The new
encoding does not need to be scanned for pointers and can be moved all at once
when compacting. The new encoding is also a bit more compact.

In the new encoding, the former heap addresses, which are heap-relative offsets,
are now stored as signature-relative offsets instead. The new encoding is also
4-byte aligned instead of 8-byte aligned, so we spend less space storing the
offsets themselves. Finally, we can elide the headers we stored for each object
in the heap, saving further space.

My motivation here is not purely performance, however. In follow-up work, I plan
to make the shared heap a bit more complicated. By offloading much of the
current complexity of the heap to this new encoding, that future work is quite a
bit simpler.

Reviewed By: mroch

Differential Revision: D32638119

fbshipit-source-id: 1810da885593344eb8460c71770679ee154488ca