Add non-batch oldify API to SharedMem
Summary:
These batch functions just iterate over a set calling oldify on each element. We
might as well expose the simpler function.

Adding this because in a follow-up I want to oldify elements from a list without
constructing a set. Ultimately I am going to remove oldify/revive entirely, so
this is temporary.

Reviewed By: mroch

Differential Revision: D32710968

fbshipit-source-id: d2f86f5d0249243f03ed6830bb64edd6b8adff9b