Simplify Commit_modules_mutator
Summary:
1. Replace `Modulename.Set.t ref option ref` with just `Modulename.Set.t ref`

There's no reason for multiple refs here. There's just one set of currently
oldified modules. The outer ref is set to the inner ref when the mutator is
created, the inner ref is mutated as modules change, and the outer ref is
cleared at the end of the transaction when the mutator becomes garbage.

2. Remove unnecessary conversion of the to_replace list to a set. We can just
add each element to the currently oldified modules ref and then iterate over the
list to oldify each hash table key.

Reviewed By: mroch

Differential Revision: D32711004

fbshipit-source-id: def2b7f406b054ab5c3375f1a611fe2ba4e52261