remove force-recheck --profile from tests
Summary:
`flow force-recheck --profile` waits for the recheck. it was being used in the `direct_dependent_files_cache` test so that we can tail the log and see if the resolved requires changed.

`flow status` also waits for the recheck, and doesn't cause one. so instead of using `--profile` (and then ignoring the profiling info), we can just use the non-blocking `force-recheck` and block using `status`, and the "Resolved requires are (un)changed" line we're interested in at the end of the logs will still be the last "Resolved requires" line.

the `OUT=$()` nonsense I'm doing here is just to avoid changing the test output in this diff, to prove that this is a no-op.

Reviewed By: samwgoldman

Differential Revision: D32734748

fbshipit-source-id: 646c6fefaaf56e8d87cf9c7ba3ea6c3de2241ddb