remove Lwt when handling ephemeral commands immediately
Summary:
as with D32731450, this removes Lwt when handling "immediate" ephemeral commands (we respond to them as soon as we read them, without putting them in the queue).

it was already the case that these didn't do anything yieldy, so this shouldn't change any behavior, but will ensure we don't do anything yieldy in the future.

it also achieves the end goal of making `ServerMonitorListener.handle_message` synchronous. this ensures that we read messages from the monitor socket without accidentally yielding to another thread. we don't want to yield because we want to eagerly process all messages from the monitor before handling them. and we want to do that because the monitor may have sent cancellation requests, multiple file watcher notifications, etc, and we don't want to leave them sitting in the socket while doing work that will be invalidated.

Reviewed By: samwgoldman

Differential Revision: D32736697

fbshipit-source-id: 5f08195bc8c91e9a5a65706d01b00bcf52bed180