Make refinement keys encode locations
Summary:
In order to fix the `failwith` for Projections, we need to be able to look up the type of a projection at a position. The first step to enable that behavior is to record the location of where the Projection occurred. This can be done by associating a location with the refinement key of the projection.

This diff modifies the types so that we have that ability. The next diff uses these new types to add a location to the Projection variant. The diff after that one uses these locations to create types.

No test number change here

Reviewed By: mvitousek

Differential Revision: D31802301

fbshipit-source-id: 2bfe449efd773bd2a00d73343034d384c77e8067