Improve behavior of projection merging after loops
Summary:
Since we do not merge heap refinements if only one branch has one, we need to take special care to ensure that we apply heap refinements at the end of a loop.

See the large inline comment for a detailed explanation of these subtle changes.

This diff has no effect on the test pass rate. So why is it in this stack?

Note the `Projection` created in the `heap_default` argument that I've removed here. There is no good location to give that projection with the information available at that point in the code. These changes give it the correct location by construction.

Reviewed By: mvitousek

Differential Revision: D31802086

fbshipit-source-id: 182387e2fbe7df62e9d904d063b0a6de10b26edf