read react at jsx_element position
Summary:
Statement.ml is going to query the environment for a `React` identifier (or whatever else is specified using the Huxpro pragma) on every jsx element. This is tricky because there is no actual location for the identifier.

There's two cases we care about. If the jsx function name is just an identifier, then statement.ml [looks up that identifier at the location of the element]( https://www.internalfb.com/code/fbsource/[8b27f9943c8adb4bb5d1091da190c45e2925a55e]/fbcode/flow/src/typing/statement.ml?lines=6501-6510). Otherwise, statement.ml traverses the Jsx pragma expression in the comment of the file, so we do the same exact thing in the EnvBuilder.

We need to also traverse this expression in the ssa_builder, since that tells the env builder which names may be unbound. Since the names in the jsx expression may be unbound, we record them all.

Finally, the Jsx pragma contains ALoc.ts, so we can no longer functorize the EnvBuilder over L. Instead, we concretely use ALoc.ts in the EnvBuilder.

See the adding test for an example of a non-identifier expression used for jsx. Notice that we emit a `cannot resolve name` error, since `call` is not in scope at the point of the jsx syntax.

Test results

`Passed: 187, Failed: 282, Skipped: 273, Errored: 0` (for now, heap refinements are still disabled in this diff)

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D31696160

fbshipit-source-id: f3f8668122f7d79f45e8dcfe5062437ea06b7691