Use heap refinements reported by EnvBuilder
Summary:
This diff uses the heap refinements reported by the EnvBuilder instead of predicates of condition. This causes new crashes because Projections are still not yet implemented. The diffs following this one in the stack bring the errored tests back down to 0. This diff enables the heap refinements even though they can cause crashes (if env_mode=ssa is enabled, which it is not in any project in the universe-- I'm confident about that) so that the incremental improvements from the diffs coming after it show a clearer impact on the test case error reduction.

`Passed: 182, Failed: 267, Skipped: 273, Errored: 20`

Reviewed By: mvitousek

Differential Revision: D31736402

fbshipit-source-id: fa490e868f3743d859f7bdc7467c929a63d174bd