Restore discriminant reads after traversing case expression
Summary:
When reading case expressions in a switch statement, the env-builder will compare the discriminant against each expression. This causes the expression to be re-read, and it can contain an unrefined Projection that is created by the refinement generated from the case expression.

This can cause a projection at the location of the discriminant to be the val entry at that location, which means the new-env will try to look up its type in the map before it is traversed. To avoid this, we restore the original state of the read at the discriminant's location after each case is evaluated.

After an update to D31803093, this behavior was no longer causing crashes, but it was still wrong.

Reviewed By: mvitousek

Differential Revision: D31803288

fbshipit-source-id: f674fad62f87289d2f3b136dcc181359347f924b