Drop file sig heap cache
Summary:
The file sig data we store in shared memory contains information about requires
and exports. The exports data is unused, and we use the new type sig data
instead.

However, requires are still used for a few things:
* We read the requires so we can resolve them
* We read the requires and combine with resolved requires when constructing the
  dependency graph
* We read the requires and combine with resolved requires to respond to
  get-imports commands
* We read the requires and combine with resolved requires during check to locate
  the dependency files so we can load their signatures

In all of the above cases, we read the file sig for a given file once, so
caching has no benefit.

Reviewed By: mroch

Differential Revision: D32776825

fbshipit-source-id: 54e33e1df766c511494391c2a9dccdd2a40394eb