Tweak caching for AST heap
Summary:
Before this diff, we used SharedMem.WithCache to create the AST heap. The
WithCache functor embeds a LocalCache inside the heap itself. The cache is
populated on add and get, cleared on oldify and remove.

(Note that get_old queries will never hit the cache, by design, so IDE commands
may not benefit while a recheck is ongoing.)

A few problems with this:
* We don't really want to add to the cache on adds, which happens during parse,
  because we never read the AST back out. Caching here just means that the ASTs
  can't be garbage collected, wasting precious memory.
* The SharedMem.WithCache code ends up caching the "relative loc" ASTs, which we
  still need to convert into ASTs using Loc.t. This means we end up copying a
  lot of the spine of the AST despite caching.

This diff hoists the cache out of the heap, and in doing so solves both the
problems above:
* We just don't add to the cache when adding ASTs to the heap. Simple.
* On reads, we add to the cache after "decompactifying" the ASTs, so cache hits
  don't need to repeat that work.

Finally, this diff enables us to move the ASTs into the existing TypeSig heap,
similar to how we store docblock and aloc tables. This will free up space in the
hash table, and moves us one small step closer to a simpler shared memory
architeture.

Reviewed By: mroch

Differential Revision: D32776263

fbshipit-source-id: 0cddcc008d21df3df2e8c700d283f6678b3cde03