optimize enums during union check
Summary:
When processing UnionT ~> UnionT constraints (in `union_optimization_guard`),
before going into point-wise comparison, we first check for the case where the two
unions are enums and so the subtyping check becomes a subset check. For this
special-case to work the unions need to have been optimized first as enums. However,
there is no guarantee that these enums are going to be optimized and so we might be
missing on a crucial shortcut.

This diff makes sure that enums have been optimized in the above case, so that we
don't miss a potential opportunity for a fast enum ~> enum fast. Note that we do not
call `UnionRep.optimize` at that point, but instead export an `optimize_enum_only`
function, that only optimizes for the enum the case. There's a couple of reasons for this:
(i) this function is not using any of the other optimized union forms, and (ii) calling the
more general `optimize` causes the test added in the previous diff to not terminate.

Changelog: [perf] Improve performance of checking union types by optimizing enums more aggressively

Reviewed By: samwgoldman

Differential Revision: D33013513

fbshipit-source-id: 6c3f0e22598de4e611625c4e4322374a31eb85c7