add typeapp termination test
Summary:
An earlier version of the next diff (D33013513) would cause a "recursion limit
exceeded" error. The test added here is a mini-repro of that error.

The change that caused this behavior was calling `UnionRep.optimize` in
`union_optimization_guard`. While calling this optimization shouldn't make a
difference in the checker, some of the constraints in subtyping_kit.ml are handled
differently based on whether the union has been optimized.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D33040733

fbshipit-source-id: c0dc1c9918d49fddb242109d5cd21588a3968c35