add test for unexpected change in dependency during recheck
Summary:
when rechecking in lazy mode + saved state, some files might not have been parsed yet. `ensure_parsed` checks for this as well as any unexpected changes (a race with the file watcher).

when a file hasn't been parsed, we pretend we got a file system update about it and restart the recheck. this reparses, merges, and checks the file and its dependents.

I intend to tweak this behavior, but adding tests of the current behavior so that followup diffs can show their changes.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D32952655

fbshipit-source-id: b4af9c9998606dfc63762747b8782f5eb6418708