Enable type lookups in new_env
Summary:
This diff redirects type lookups into the output from the name_resolver. The changes outside of the new_env.ml file were to fix crashes introduced by this change. These crashes were querying the declaration site of a type alias for its read type, but should instead be looking at its declared type.

New results:
`Passed: 174, Failed: 300, Skipped: 265, Errored: 0`

Note that this is a regression from the previous diff. That's ok, though. We'd prefer to handle the types in the new env and fix the issues created by it over using the old env for types.

Reviewed By: panagosg7

Differential Revision: D32317203

fbshipit-source-id: 1223d7f7dc2fd8abad23a2e51f45c7b41aa4b79b