Implement switch exhaustiveness checking in the name_resolver
Summary:
We say that a switch is exhaustive is conjuncting the negation of the refinements introduced by each case cases a type to be filtered to `empty`. [Statement.ml currently does that here](https://www.internalfb.com/code/fbsource/[fdb7942d1c77a2796a65d499f126a17e9103f90b]/fbcode/flow/src/typing/statement.ml?lines=1487-1508). This approach requires special support in the name_resolver, since that value is synthetic and may never actually appear in the source code.

The name_resolver will now associate a "discriminant with all negated refinements" at the location of the switch statement (implemented in this diff), which will let the new_env query those values to determine if a switch is exhaustive (implemented in the next diff).

Note that we just add this Val to the `values` map. We could hypothetically create a new mapping for this (indeed, my original implementation did), but since there is no risk of keys clashing (since expressions cannot have the same loc as an entire swith statement) that extra map seems unneccessary.

Reviewed By: panagosg7

Differential Revision: D32364334

fbshipit-source-id: ea8b566c448d0d6528074cd9939ecb26801ed84c