Correctly init imports
Summary:
Imports query the var_declared_type for a tvar and then unify the imported type with that tvar.

This does not play well with the new-env. The provider analysis reports value imports as providers for themselves and no providers for type imports. But it would not be reasonable for the provider analysis to be able to provide a real type for the imports.

To ensure that imports are correctly populated, we let each env implementation customize it's action via the newly introduced `init_import` function.

In the new_env, we model this as an assignment. In the current (future: old) env, we continue to query the var's general type and then unify the provided type with it.

`Passed: 274, Failed: 202, Skipped: 269, Errored: 0`

Reviewed By: panagosg7

Differential Revision: D32143853

fbshipit-source-id: 60e7e0fd8437054fb8349b2290a5d094b7a111b0