Fix local_scope_entry_exists in new_env.ml
Summary:
This function used to return true if the env_builder had an entry at the location queried. But the env_builder should always have those entries!

Instead of just querying the loc_env to see if a value is present, we should also determine if that value is Global or not. This diff implements that behavior, which gets rid of several errors related to `require`, `module`, `exports`, etc.

New stats:
`Passed: 239, Failed: 236, Skipped: 269, Errored: 0`

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D32327307

fbshipit-source-id: 21e955d20179d467a87c0ed0de55a108ba2e9586