add unexpected updates as dependencies
Summary:
when rechecking in lazy mode + saved state, `ensure_parsed` makes sure that all relevant files have been parsed, or else we'd be unable to merge them. if a file is unexpectedly changed (e.g. a race with the file watcher), we can't just parse it because it could impact the dependency graph. today, we just pretend that we got a file system event for that file and restart the recheck to include it. this parses, merges and checks the file, and also all of its dependents.

this is correct, but expensive. it's a problem for the IDE because sometimes we're trying to quickly *merge* some dependencies to unblock the IDE request (in `experimental.prioritize_dependency_checks`), find them unexpectedly changed and end up *checking* all *dependents* of these changed dependencies.

instead, we need to be lazier. here, I introduce the concept of a "dependency" update, where we only do the necessary work to use the changed file as a dependency: fix up the dependency graph and re-merge it, but don't re-check it or any of its dependents.

when we eventually receive the file watcher update, it will be "upgraded" to a normal update, which will cause the file and its dependents to be rechecked like before.

Changelog: [fix] Improve performance of IDE requests when using lazy mode and saved state

Reviewed By: samwgoldman

Differential Revision: D32952654

fbshipit-source-id: c08cee4553ba188aaac0c265d5175a8e5355e970