Re-organize shared memory API by data type
Summary:
Before this change, the API was arranged by function -- that is, functions to
write to shared memory, functions to read, etc.

This diff changes the API to be arranged by data type. Functions for dealing
with strings -- sizing, reading, writing -- are placed next to each other.

Over the last few months, I've found that I often need to change functions for a
given data type together. For example, adding a new field affects the size, how
the value is written, and how the value is read back out.

Special mention to nmote who advocated for this layout originally. He was right
and I was wrong.

Reviewed By: mroch

Differential Revision: D33098686

fbshipit-source-id: 9cc916d4efb27f5bcb0b6b07b1e5243dcbb673b9