Move ASTs to the new heap
Summary:
Before this diff, we stored ASTs in shared memory through the ASTHeap, which
comes from the SharedMem.NoCache functor. For each AST, we added an entry to the
shared hash table.

This diff removes the ASTHeap and instead stores the address of the
serialized+compressed AST in the checked_file heap object.

Compression is important for storing ASTs. Most of the complexity here is
implementing the compression scheme from hh_shared.

Reviewed By: mroch

Differential Revision: D33100301

fbshipit-source-id: f86fcb709f347bbae6ee2bae6654ab3e5ddb0766