use CheckedSet for recheck updates
Summary:
this diff is a rather mechanical change from using `FilenameSet` to `CheckedSet` for `updates`. `updates` is the set of files we've gotten change notifications about.

I use `CheckedSet.all` everywhere, which makes it equivalent to the previous `FilenameSet`. there is no behavior change here.

this gets the refactoring out of the way of the next diff.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D32952657

fbshipit-source-id: 33343ab889ecddc20c31471e7baba883e1e11d76