Support for this tparam lookups
Summary:
`this` type params for classes are stored using the same location for the class. This is obviously an issue for the new-env because that's where we store the type of the class!

In order to support queries for `this` params, we create a separate mapping in the new_env to respond to queries to `this` type param lookups.

`Passed: 339, Failed: 137, Skipped: 269, Errored: 0`

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D32366997

fbshipit-source-id: 6361e5850b44ae39c5ea1b99992c061e3e0d7bdf