Let the parse phase discover deleted files
Summary:
A file update might be a new file, changed file, or deleted file. The server
just receives a collection of undifferentiated updated file names.

Before this change, we first split the list of updates into modified and deleted
by checking the file system. However, the files might be deleted after we check
and before we parse, so the parse phase needs to handle missing files too.

After this change, we do not bother checking the file system first, and pass the
entire list of updated file paths to the parse phase. Files which can not be
found are accumulated and returned back to the caller.

Reviewed By: mroch

Differential Revision: D33204405

fbshipit-source-id: 67858bf444a4fcaf6648f063ec8628247cadf1c4